Break execution-plans into more modules
